+++ java/src/main/java/com/google/j2cl/samples/wasm/MatrixOperations.java
@@ -4,15 +4,15 @@ import static com.google.j2cl.samples.wasm.Matrix.*;
 public class MatrixOperations {
     public static String doMatrixOperations(int matrixSize) {
         Matrix.matrixSize = matrixSize != 0 ? matrixSize : 200;
-        long[][] matrix1 = getEmptyMatrix();
-        long[][] matrix2 = getEmptyMatrix();
+        double[][] matrix1 = getEmptyMatrix();
+        double[][] matrix2 = getEmptyMatrix();
         fillSequentialMatrix(matrix1);
         fillSequentialMatrix(matrix2);
-        long total_sum = 0L;
+        double total_sum = 0L;
         for (int i = 0; i < 100; i++) {
-            long[][] result = multiplyMatrices(matrix1, matrix2, i);
+            double[][] result = multiplyMatrices(matrix1, matrix2, i);
             total_sum += sumOfMatrixElements(result);
         }
-        return Long.toString(total_sum);
+        return Double.toString(total_sum);
     }
 }
